*refactoring of ConvertScatterElementsNode.
